fix(TableList): Fixed duplicate code, move SingleandDoubleClick function to separate file and remove the css style for cursor
