+++ app/styles/_table-list.scss
@@ -188,12 +188,6 @@ $table-list-palette: $sd-basic-palette;
         }
     }
 }
-.table-list__item-content-block--title {
-    * {
-        cursor: text;
-    }
-}
-
 
 .table-list__slide-in-actions,
 .table-list__static-actions,
